Backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
anyio==4.4.0
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
bce-python-sdk==0.9.17
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
dnspython==2.6.1
docutils==0.21.2
erniebot==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.3.0
multidict==6.0.5
networkx==3.3
numpy==1.24.4
open-python==1.1.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1.post120
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
protobuf==3.20.2
psutil==6.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyMuPDF==1.24.6
PyMuPDFb==1.24.6
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
rarfile==4.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.31
termcolor==2.4.0
tifffile==2024.6.18
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visualdl==2.5.3
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
